Add 'glyph' and 'rune' subcommands
Examples:

    $ urb rune ^-
    ^-  [model value]
    typecast by mold
    aka: `foo`bar

    $ urb glyph "_+>$&*"
    cab lus ban bus pad tar

This patch also refactors some of the cli parser stuff. One important change is
the top-level commands have been moved into the `run` subcommand. This was
done to make the cli parser code simpler, and I think it's the right way to do
it, but I'm open to other suggestions.